Refactor sources (2/x)
